Started using yfinance, changed precision on price data and updated tests
Due to an issue with pandas-datareader and yahoo, started using yfinance. This produced slightly different adjusted values, which broke the unit test.

To counter the large precision differences from the price adjustment, decided to round the values to 2 decimal places. This produced a number of other test case issues, but was quickly resolved by reducing the precision of the test values to 2 decimal places from 4.